+++ packages/shared/src/modules/default_modules/router/components/uiLayerRouter.tsx
@@ -21,9 +21,10 @@ import {Fragment} from "react/jsx-dev-runtime";
 import {useEffect} from "react";
 import {Route, Routes, useLocation, useNavigate} from "react-router";
 import {generateHash} from "../../../../utils/hash";
-import {useDefaultSelector} from "../../../module_orchestration/moduleDefaults";
+import {useSelector} from "react-redux";
 import {BrowserRouter} from "react-router-dom";
 import {UILayerRouterProps} from "../../../../types/modules/router/routerModule";
+import type {DefaultRootState} from "../../../../types/modules/moduleDefaultTypes";
 
 export const UILayerRouter = (props: UILayerRouterProps) => {
   const enabledRoutes = props.routes.filter((route) => !route.disabled);
@@ -61,7 +62,9 @@ interface RedirectorProps {
  * @constructor
  */
 const Redirector = (props: RedirectorProps) => {
-  const {hasAuthenticated} = useDefaultSelector((state) => state.auth);
+  const {hasAuthenticated} = useSelector(
+    (state: DefaultRootState) => state.auth,
+  );
 
   const disableLogin = props.disableLogin;