+++ packages/shared/src/modules/default_modules/ui/header/components/settingsMenu.tsx
@@ -23,8 +23,9 @@ import {ColorSettingsContext} from "../../../../../contexts/colorSettingsContext
 import {useModule} from "../../../../../contexts/moduleContext";
 import {MandatoryModuleNames} from "../../../../../constants/moduleNames";
 import {useModuleTranslation} from "../../../../hooks/useModuleTranslation";
-import {useDefaultSelector} from "../../../../module_orchestration/moduleDefaults";
+import {useSelector} from "react-redux";
 import {LangResources} from "../../../../../types/modules/internationalization/internationalizerModule";
+import type {DefaultRootState} from "../../../../../types/modules/moduleDefaultTypes";
 
 // ##############################################
 // Notice: The enclosed imports are copied from 'primereact/menuitem/MenuItem' as the path could not be resolved by the gitlab builder
@@ -91,8 +92,9 @@ export const SettingsMenu = React.forwardRef<ContextMenu, Props>(
     const basicOptions: MenuItem[] = [];
     let notFallbackLang = false;
 
-    const activeLang = useDefaultSelector(
-      (state) => state[MandatoryModuleNames.Internationalizer].activeLang,
+    const activeLang = useSelector(
+      (state: DefaultRootState) =>
+        state[MandatoryModuleNames.Internationalizer].activeLang,
     );
 
     if (!props.menuOptions?.hideLanguageSelection) {