+++ packages/shared/src/modules/default_modules/ui/header/components/userMenu.tsx
@@ -21,7 +21,8 @@ import {ContextMenu} from "primereact/contextmenu";
 import {MenuItem} from "./settingsMenu";
 import {useModule} from "../../../../../contexts/moduleContext";
 import {MandatoryModuleNames} from "../../../../../constants/moduleNames";
-import {useDefaultDispatch} from "../../../../module_orchestration/moduleDefaults";
+import {useDispatch} from "react-redux";
+import type {DefaultThunkDispatch} from "../../../../../types/modules/moduleDefaultTypes";
 
 export interface Props {
   hideMenu: (e: React.KeyboardEvent) => void;
@@ -37,7 +38,7 @@ export interface UserMenuOptions {
 export const UserMenu = React.forwardRef<ContextMenu, Props>((props, ref) => {
   const authModule = useModule(MandatoryModuleNames.Authenticator);
 
-  const dispatch = useDefaultDispatch();
+  const dispatch: DefaultThunkDispatch = useDispatch();
 
   const basicOptions: MenuItem[] = [];